changed class names
Removed firstLevel and facetLevel class names as they don't sound like
they fit the scope of this plugin. new class names will be
tbTree-level-0, tbTree-level-1, etc depending on the depth of the
branches. For all other classes that are used in the plugin I have
prepended "tbTree-" as well. I have also removed any unused classes that
don't aide in the functionality of the plugin or relate to any type of
styling like bootstrap.